requirements.txt
twisted-web>=12.3.0
twisted-core>=12.3.0
nltk>=3.0.3
